renaming files with proper naming conventions.
